Encapsulate processes dealing with directing the lifecycle of a ruleset.
