+++ HouseRules_Core/HouseRules_Core.csproj
@@ -48,7 +48,7 @@
     </ItemGroup>
     <ItemGroup>
         <Compile Include="BoardSyncer.cs" />
-        <Compile Include="ModPatcher.cs" />
+        <Compile Include="LifecycleDirector.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="HR.cs" />
         <Compile Include="CoreMod.cs" />
